requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=1.6
Babel>=1.3
oslo.i18n>=1.5.0
oslo.utils!=2.6.0,>=2.4.0
python-keystoneclient
python-ironicclient
python-oneviewclient
requests